backend/products/requirements.txt
fastapi
uvicorn[standard]
psycopg2-binary
sqlalchemy
kafka-python
grpcio
grpcio-tools
python-jose[cryptography]